+++ packages/seedcord/src/bot/injectors/EmojiInjector.ts
@@ -1,6 +1,5 @@
+import { Logger } from '@seedcord/services';
 import chalk from 'chalk';
-
-import { Logger } from '../../services/Logger';
 
 import type { Core } from '../../interfaces/Core';
 
@@ -28,6 +27,8 @@ export class EmojiInjector {
         configEmojis[key] = `<${emoji.identifier}>`;
 
         foundCount++;
+
+        this.logger.debug(`${chalk.bold.green('Found')}: ${chalk.magenta.bold(emojiName)} (${emoji.id})`);
       }
     });