+++ packages/seedcord/src/index.ts
@@ -12,15 +12,7 @@ export * from './bot/decorators/EventCatchable';
 export * from './bot/decorators/EventRegisterable';
 export * from './bot/decorators/InteractionConfigurable';
 
-// Services exports
-export * from './services/Logger';
-export * from './services/Lifecycle/CoordinatedStartup';
-export * from './services/Lifecycle/CoordinatedShutdown';
-export * from './services/HealthCheck';
-export * from './services/CooldownManager';
-
 // Library exports
-export * from './library/Globals';
 export * from './library/Helpers';
 
 // Interfaces exports
@@ -40,3 +32,12 @@ export type * from './effects/types/Effects';
 
 // Errors
 export * from './bot/errors/Database';
+
+// Export Services
+export * from '@seedcord/services';
+
+// Export Types
+export type * from '@seedcord/types';
+
+// Export Utils
+export * from '@seedcord/utils';