+++ packages/seedcord/src/interfaces/Components.ts
@@ -2,24 +2,29 @@ import {
   ActionRowBuilder,
   ButtonBuilder,
   ChannelSelectMenuBuilder,
+  ContainerBuilder,
   ContextMenuCommandBuilder,
   EmbedBuilder,
+  FileBuilder,
   InteractionContextType,
+  MediaGalleryBuilder,
   MentionableSelectMenuBuilder,
   ModalBuilder,
   RoleSelectMenuBuilder,
+  SectionBuilder,
+  SeparatorBuilder,
   SlashCommandBuilder,
   SlashCommandSubcommandBuilder,
   SlashCommandSubcommandGroupBuilder,
   StringSelectMenuBuilder,
   StringSelectMenuOptionBuilder,
+  TextDisplayBuilder,
   TextInputBuilder,
   UserSelectMenuBuilder
 } from 'discord.js';
+import { Envapt } from 'envapt';
 
-import { Globals } from '../library/Globals';
-
-import type { ModalActionRowComponentBuilder } from 'discord.js';
+import type { ColorResolvable, ModalActionRowComponentBuilder } from 'discord.js';
 
 const BuilderTypes = {
   command: SlashCommandBuilder,
@@ -34,7 +39,13 @@ const BuilderTypes = {
   modal: ModalBuilder,
   context_menu: ContextMenuCommandBuilder,
   subcommand: SlashCommandSubcommandBuilder,
-  group: SlashCommandSubcommandGroupBuilder
+  group: SlashCommandSubcommandGroupBuilder,
+  container: ContainerBuilder,
+  text_display: TextDisplayBuilder,
+  file: FileBuilder,
+  media: MediaGalleryBuilder,
+  section: SectionBuilder,
+  separator: SeparatorBuilder
 };
 
 const RowTypes: {
@@ -132,12 +143,15 @@ abstract class BaseComponent<TComponent> {
 export abstract class BuilderComponent<BuilderKey extends BuilderType> extends BaseComponent<
   InstantiatedBuilder<BuilderKey>
 > {
+  @Envapt<ColorResolvable>('DEFAULT_BOT_COLOR', { fallback: 'Default' })
+  declare private readonly botColor: ColorResolvable;
+
   protected constructor(type: BuilderKey) {
     const ComponentClass = BuilderTypes[type] as unknown;
     super(ComponentClass as new () => InstantiatedBuilder<BuilderKey>);
 
     // Override in builders
-    if (this.instance instanceof EmbedBuilder) this.instance.setColor(Globals.botColor);
+    if (this.instance instanceof EmbedBuilder) this.instance.setColor(this.botColor);
 
     // Override in builders
     if (this.instance instanceof SlashCommandBuilder || this.instance instanceof ContextMenuCommandBuilder) {